requirements.txt
torchvision==0.7.0
numpy==1.19.2
torch==1.6.0
scikit_image==0.17.2
tqdm==4.47.0
scipy==1.4.1
Pillow==8.2.0
skimage==0.0
